requirements.txt
playwright
requests
pandas
openpyxl
xlrd
pywin32
python-dotenv
rich
